requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
blis==0.7.9
boto3==1.26.104
botocore==1.29.104
bs4==0.0.1
cached-property==1.5.2
catalogue==2.0.8
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
chex==0.1.7
cleantext==1.1.4
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
confection==0.0.4
contextlib2==21.6.0
contourpy==1.1.0
cramjam==2.6.2
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==4.4.2
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dm-tree==0.1.8
dnspython==2.4.2
dython==0.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
etils==1.2.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.1
fastparquet==2023.7.0
ffmpeg-python==0.2.0
filelock==3.11.0
flax==0.6.9
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
google-pasta==0.2.0
googletrans==3.1.0a0
GPUtil==1.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jax==0.4.8
jaxlib==0.4.7
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.128
langcodes==3.3.0
libretranslatepy==2.1.1
llvmlite==0.40.1rc1
loguru==0.7.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.7
numba==0.57.0
numpy==1.23.5
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.2.1
outcome==1.2.0
packaging==23.0
pandas==1.5.1
parso==0.8.3
pathos==0.3.0
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.2.0
plotly==5.15.0
pox==0.3.2
ppft==1.7.6.6
pprintpp==0.4.0
preshed==3.0.8
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==3.20.3
protobuf3-to-dict==0.1.5
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.7
pydub==0.25.1
Pygments==2.14.0
pyodbc==4.0.38
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.28.2
rfc3986==1.5.0
rich==13.3.4
s3transfer==0.6.0
sagemaker==2.151.0
schema==0.7.5
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tblib==1.7.0
tenacity==8.2.2
tensorstore==0.1.36
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
trio==0.22.0
trio-websocket==0.10.2
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
ua-parser==0.16.1
urllib3==1.26.15
user-agents==2.2.0
uvicorn==0.22.0
wasabi==1.1.1
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
